# Use max heap for building query result