It requires more allocations (not constant anymore), but is 1) faster and 2) requires much *smaller* allocations, so overall a win.

```
goos: linux
goarch: amd64
pkg: github.com/philippgille/chromem-go
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
                                    │    after2     │               after3                │
                                    │    sec/op     │    sec/op     vs base               │
Collection_Query_NoContent_100-8      106.32µ ±  2%   90.79µ ±  2%  -14.61% (p=0.002 n=6)
Collection_Query_NoContent_1000-8      490.7µ ±  1%   518.8µ ±  1%   +5.74% (p=0.002 n=6)
Collection_Query_NoContent_5000-8      2.704m ±  5%   2.144m ±  1%  -20.72% (p=0.002 n=6)
Collection_Query_NoContent_25000-8    13.171m ± 11%   9.947m ±  2%  -24.48% (p=0.002 n=6)
Collection_Query_NoContent_100000-8    51.86m ± 13%   39.75m ±  1%  -23.34% (p=0.002 n=6)
Collection_Query_100-8                106.77µ ±  0%   90.99µ ±  1%  -14.78% (p=0.002 n=6)
Collection_Query_1000-8                489.7µ ±  0%   595.2µ ± 13%  +21.55% (p=0.002 n=6)
Collection_Query_5000-8                2.704m ±  6%   2.556m ±  1%   -5.47% (p=0.002 n=6)
Collection_Query_25000-8               13.05m ±  2%   11.66m ±  1%  -10.65% (p=0.002 n=6)
Collection_Query_100000-8              52.07m ±  5%   39.70m ± 12%  -23.76% (p=0.002 n=6)
geomean                                2.492m         2.192m        -12.07%

                                    │    after2     │               after3                │
                                    │     B/op      │     B/op      vs base               │
Collection_Query_NoContent_100-8       6.235Ki ± 0%   5.030Ki ± 0%  -19.32% (p=0.002 n=6)
Collection_Query_NoContent_1000-8      34.74Ki ± 0%   13.24Ki ± 0%  -61.88% (p=0.002 n=6)
Collection_Query_NoContent_5000-8     162.74Ki ± 0%   45.99Ki ± 0%  -71.74% (p=0.002 n=6)
Collection_Query_NoContent_25000-8     794.7Ki ± 0%   206.7Ki ± 0%  -73.99% (p=0.002 n=6)
Collection_Query_NoContent_100000-8   3130.7Ki ± 0%   791.4Ki ± 0%  -74.72% (p=0.002 n=6)
Collection_Query_100-8                 6.234Ki ± 0%   5.033Ki ± 0%  -19.27% (p=0.002 n=6)
Collection_Query_1000-8                34.74Ki ± 0%   13.25Ki ± 0%  -61.87% (p=0.002 n=6)
Collection_Query_5000-8               162.73Ki ± 0%   46.04Ki ± 0%  -71.71% (p=0.002 n=6)
Collection_Query_25000-8               794.7Ki ± 0%   206.8Ki ± 0%  -73.98% (p=0.002 n=6)
Collection_Query_100000-8             3130.8Ki ± 0%   791.4Ki ± 0%  -74.72% (p=0.002 n=6)
geomean                                154.4Ki        54.97Ki       -64.40%

                                    │   after2   │               after3                │
                                    │ allocs/op  │  allocs/op   vs base                │
Collection_Query_NoContent_100-8      41.00 ± 0%    94.00 ± 1%  +129.27% (p=0.002 n=6)
Collection_Query_NoContent_1000-8     41.00 ± 0%   140.50 ± 0%  +242.68% (p=0.002 n=6)
Collection_Query_NoContent_5000-8     41.00 ± 0%   172.00 ± 1%  +319.51% (p=0.002 n=6)
Collection_Query_NoContent_25000-8    41.00 ± 0%   204.00 ± 1%  +397.56% (p=0.002 n=6)
Collection_Query_NoContent_100000-8   41.00 ± 0%   232.00 ± 3%  +465.85% (p=0.002 n=6)
Collection_Query_100-8                41.00 ± 0%    94.50 ± 1%  +130.49% (p=0.002 n=6)
Collection_Query_1000-8               41.00 ± 0%   141.00 ± 1%  +243.90% (p=0.002 n=6)
Collection_Query_5000-8               41.00 ± 0%   174.50 ± 2%  +325.61% (p=0.002 n=6)
Collection_Query_25000-8              41.00 ± 0%   205.50 ± 2%  +401.22% (p=0.002 n=6)
Collection_Query_100000-8             41.50 ± 1%   233.00 ± 1%  +461.45% (p=0.002 n=6)
geomean                               41.05         161.4       +293.09%
```